refactor(suite): reorganize and rename files in send folder
