+++ packages/suite/src/views/wallet/send/Options/EthereumOptions/EthereumOptions.tsx
@@ -4,7 +4,7 @@ import { FormOptions } from '@suite-common/wallet-types';
 import { Button, Tooltip } from '@trezor/components';
 import { Translation } from 'src/components/suite';
 import { useSendFormContext } from 'src/hooks/wallet';
-import { Data } from './Data';
+import { EthereumData } from './EthereumData';
 import { OnOffSwitcher } from '../OnOffSwitcher';
 
 const Wrapper = styled.div`
@@ -73,7 +73,7 @@ export const EthereumOptions = () => {
                 </Tooltip>
             </Left>
 
-            {dataEnabled && <Data close={toggleData} />}
+            {dataEnabled && <EthereumData close={toggleData} />}
         </Wrapper>
     );
 };
