+++ packages/suite/src/views/wallet/send/Options/OnOffSwitcher.tsx
