+++ packages/suite/src/views/wallet/send/Options/Options.tsx
@@ -3,7 +3,7 @@ import { useSendFormContext } from 'src/hooks/wallet';
 import { BitcoinOptions } from './BitcoinOptions/BitcoinOptions';
 import { EthereumOptions } from './EthereumOptions/EthereumOptions';
 import { RippleOptions } from './RippleOptions/RippleOptions';
-import { CardanoOptions } from './CardanoOptions/CardanoOptions';
+import { CardanoOptions } from './CardanoOptions';
 
 export const Options = () => {
     const {
