+++ packages/suite/src/views/wallet/send/Outputs/Amount/SendMaxSwitch.tsx
