+++ packages/suite/src/views/wallet/send/Outputs/Amount/TokenSelect.tsx
