+++ packages/suite/src/views/wallet/send/TotalSent/TotalSent.tsx
